requirements.txt
requests==2.26.0
tweepy==4.4.0
google-api-python-client==2.33.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
oauth2client==4.1.3